settings structure change
