+++ client/src/routes/index.ts
@@ -4,3 +4,4 @@ export { Loading } from "./Loading";
 export { default as Likes } from "./home/Likes";
 export { default as Plates } from "./home/Plates";
 export { default as Tags } from "./home/Tags";
+export { default as Settings } from "./home/Settings";
